<clean> : Study 신청 로직 URL 변경
