+++ src/main/java/highsquare/hirecoder/web/controller/BoardController.java
@@ -24,7 +24,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static highsquare.hirecoder.constant.MessageConstant.MAX_APPEALMESSAGE_LENGTH;
+import static highsquare.hirecoder.constant.MessageConstant.MIN_APPEALMESSAGE_LENGTH;
 import static highsquare.hirecoder.constant.PageConstant.*;
+import static highsquare.hirecoder.constant.StudyCountConstant.STUDY_COUNT_MAX;
 
 @Controller
 @RequestMapping("/boards")
@@ -33,16 +36,14 @@ public class BoardController {
 
     private final BoardService boardService;
     private final CommentService commentService;
-
     private final LikeOnBoardService likeOnBoardService;
-
     private final StudyService studyService;
-
     private final StudyMemberRepository studyMemberRepository;
-
     private final BoardRepository boardRepository;
-
     private final TagService tagService;
+    private final StudyMemberService studyMemberService;
+    private final MessageForApplicationService messageForApplicationService;
+    private final ApplyForStudyService applyForStudyService;
 
 
     @GetMapping("/{kind}/{study_id}/{board_id}")
@@ -179,5 +180,135 @@ public class BoardController {
                 board.getPublicYn(), board.getViewCnt(), board.getLikeCnt(), board.getKind().name(), board.getCreateDate(),
                 board.getUpdateDate());
         return boardForm;
+    }
+
+    /**
+     * 일반 멤버가 스터디 신청하기 버튼 클릭시 신청테이블에 등록되는 로직
+     * 스터디에 신청할 때 사유테이블에 신청사유를 작성
+     * 우선 ScriptUtils를 이용하여 간단하게 로직 작성함(map에 오류를 넣어서 페이지로 반환시키든 리펙토링 필요)
+     */
+    @GetMapping("/RECRUIT/{studyId}/{boardId}/enroll")
+    public String checkStudyMember(@PathVariable("studyId") Long studyId,
+                                   Principal principal,
+                                   HttpServletResponse response,
+                                   Model model) throws IOException {
+
+        Long loginMemberId = Long.parseLong(principal.getName());
+
+        // 해당 스터디 존재 확인 로직
+        if (!studyService.isExistingStudy(studyId)) {
+            ScriptUtils.alertAndBackPage(response, " 해당 스터디가 존재하지 않습니다.");
+        }
+
+        if (!validateBelongedStudyCount(loginMemberId)) {
+            ScriptUtils.alert(response,"가입할 수 있는 스터디 최대 갯수를 초과하셨습니다.");
+        }
+
+        validateMemberAttendState(studyId, loginMemberId, response);
+
+        if (!validateNotEnrolled(studyId, loginMemberId)) {
+            ScriptUtils.alertAndBackPage(response, "이미 신청한 스터디입니다.");
+        }
+
+        model.addAttribute("studyId", studyId);
+
+        String studyName = studyService.getStudyNameById(studyId);
+        model.addAttribute("studyName", studyName);
+        return "form/messageForApply";
+    }
+
+    /**
+     * 신청 테이블에 대기상태로 저장됨
+     */
+    @PostMapping("/RECRUIT/{studyId}/{boardId}/enroll")
+    public String enrollApplyForStudy(@RequestParam("appealMessage") String appealMessage,
+                                      @PathVariable("studyId") Long studyId, Principal principal,
+                                      HttpServletResponse response,
+                                      Model model) throws IOException {
+
+        Long loginMemberId = Long.parseLong(principal.getName());
+
+        // 해당 스터디와 멤버가 존재하는지 확인(로그인 검증에서 멤버 존재 확인)
+        if (!studyService.isExistingStudy(studyId)) {
+            ScriptUtils.alertAndBackPage(response, " 해당 스터디가 존재하지 않습니다.");
+        }
+
+        // 스터디 멤버 테이블에서 studyId와 memberId를 검색조건으로 AttendState 상태 들고 올텐데
+        // 레코드가 존재하면 AttendState 상태에 따라 메시지를 보냄
+        // 존재하지 않을 시 신청 테이블을 생성함
+        if (validateMemberAttendState(studyId, loginMemberId, response)) {
+            // 해당 멤버의 스터디 갯수 제한 검증 로직
+            if (!validateBelongedStudyCount(loginMemberId)) {
+                ScriptUtils.alert(response,"가입할 수 있는 스터디 최대 갯수를 초과하셨습니다.");
+            }
+        }
+
+        // appealMessage 검증로직
+        if (!validMessage(appealMessage, model)) {
+            model.addAttribute("studyId", studyId);
+
+            String studyName = studyService.getStudyNameById(studyId);
+            model.addAttribute("studyName", studyName);
+            return "form/messageForApply";
+        }
+
+        // 검증 로직 종료
+
+        ApplyForStudy applyForStudy = applyForStudyService.enrollApplyForStudy(studyId, loginMemberId, AuditState.대기.name());
+        messageForApplicationService.addMessage(applyForStudy, appealMessage);
+
+        return "redirect:/study/myStudy";
+    }
+
+    private boolean validMessage(String appealMessage,Model model) {
+        if (appealMessage.length()< MIN_APPEALMESSAGE_LENGTH) {
+            model.addAttribute("invalidMessageLength", true);
+            return false;
+        }
+
+        if (appealMessage.length()> MAX_APPEALMESSAGE_LENGTH) {
+            model.addAttribute("invalidMessageLength", true);
+            model.addAttribute("rejectedMessage", appealMessage.substring(0,MAX_APPEALMESSAGE_LENGTH-1));
+            return false;
+        }
+        return true;
+    }
+
+    private boolean validateMemberAttendState(Long studyId, Long loginMemberId, HttpServletResponse response) throws IOException {
+
+        if (!studyMemberService.doesMemberBelongToStudy(studyId, loginMemberId)) {
+
+            return true;
+
+        } else {
+            AttendState attendState = studyMemberService.getAttendState(studyId, loginMemberId);
+
+            switch (attendState) {
+                case 참여 : {ScriptUtils.alertAndBackPage(response,"이미 가입하신 스터디입니다."); break; }
+                case 탈퇴 : {ScriptUtils.alertAndBackPage(response,"탈퇴한 스터디는 가입이 불가능합니다."); break; }
+                case 강퇴 : {ScriptUtils.alertAndBackPage(response,"강퇴당한 스터디는 가입이 불가능합니다."); break; }
+            }
+
+            return false;
+        }
+    }
+
+    private boolean validateBelongedStudyCount(Long loginMemberId) throws IOException {
+        int studyCount = studyMemberService.getBelongedStudyCount(loginMemberId);
+
+        if(studyCount>=STUDY_COUNT_MAX) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * 이미 신청한 상태인지 확인하는 로직
+     * @param studyId
+     * @param loginMemberId
+     */
+    private boolean validateNotEnrolled(Long studyId, Long loginMemberId) {
+        return applyForStudyService.findApplyForStudyByMemberAndStudy(studyId, loginMemberId).isEmpty();
     }
 }
