+++ src/main/java/highsquare/hirecoder/web/controller/StudyManageController.java
@@ -36,85 +36,9 @@ public class StudyManageController {
     private final ApplyForStudyService applyForStudyService;
     private final StudyService studyService;
     private final MessageForApplicationService messageForApplicationService;
-
-
-    /**
-     * 일반 멤버가 스터디 신청하기 버튼 클릭시 신청테이블에 등록되는 로직
-     * 스터디에 신청할 때 사유테이블에 신청사유를 작성
-     * 우선 ScriptUtils를 이용하여 간단하게 로직 작성함(map에 오류를 넣어서 페이지로 반환시키든 리펙토링 필요)
-     */
-    @GetMapping("/studyManage/enroll/{studyId}")
-    public String checkStudyMember(@PathVariable("studyId") Long studyId,
-                                   Principal principal,
-                                   HttpServletResponse response,
-                                   Model model) throws IOException {
-
-        Long loginMemberId = Long.parseLong(principal.getName());
-
-        // 해당 스터디 존재 확인 로직
-        if (!studyService.isExistingStudy(studyId)) {
-            ScriptUtils.alertAndBackPage(response, " 해당 스터디가 존재하지 않습니다.");
-        }
-
-        if (!validateBelongedStudyCount(loginMemberId)) {
-            ScriptUtils.alert(response,"가입할 수 있는 스터디 최대 갯수를 초과하셨습니다.");
-        }
-
-        validateMemberAttendState(studyId, loginMemberId, response);
-
-        if (!validateNotEnrolled(studyId, loginMemberId)) {
-            ScriptUtils.alertAndBackPage(response, "이미 신청한 스터디입니다.");
-        }
-
-        model.addAttribute("studyId", studyId);
-
-        String studyName = studyService.getStudyNameById(studyId);
-        model.addAttribute("studyName", studyName);
-        return "form/messageForApply";
-    }
-
-    /**
-     * 신청 테이블에 대기상태로 저장됨
-     */
-    @PostMapping("/studyManage/enroll/{studyId}")
-    public String enrollApplyForStudy(@RequestParam("appealMessage") String appealMessage,
-                                      @PathVariable("studyId") Long studyId, Principal principal,
-                                      HttpServletResponse response,
-                                      Model model) throws IOException {
-
-        Long loginMemberId = Long.parseLong(principal.getName());
-
-        // 해당 스터디와 멤버가 존재하는지 확인(로그인 검증에서 멤버 존재 확인)
-        if (!studyService.isExistingStudy(studyId)) {
-            ScriptUtils.alertAndBackPage(response, " 해당 스터디가 존재하지 않습니다.");
-        }
-
-        // 스터디 멤버 테이블에서 studyId와 memberId를 검색조건으로 AttendState 상태 들고 올텐데
-        // 레코드가 존재하면 AttendState 상태에 따라 메시지를 보냄
-        // 존재하지 않을 시 신청 테이블을 생성함
-        if (validateMemberAttendState(studyId, loginMemberId, response)) {
-            // 해당 멤버의 스터디 갯수 제한 검증 로직
-            if (!validateBelongedStudyCount(loginMemberId)) {
-                ScriptUtils.alert(response,"가입할 수 있는 스터디 최대 갯수를 초과하셨습니다.");
-            }
-        }
-
-        // appealMessage 검증로직
-        if (!validMessage(appealMessage, model)) {
-            model.addAttribute("studyId", studyId);
-
-            String studyName = studyService.getStudyNameById(studyId);
-            model.addAttribute("studyName", studyName);
-            return "form/messageForApply";
-        }
-
-        // 검증 로직 종료
-
-        ApplyForStudy applyForStudy = applyForStudyService.enrollApplyForStudy(studyId, loginMemberId, AuditState.대기.name());
-        messageForApplicationService.addMessage(applyForStudy, appealMessage);
-
-        return "redirect:/study/myStudy";
-    }
+//
+//    @GetMapping("studyManage/manager/{studyId}")
+//    public String getManagerMainPage()
 
     /**
      * 스터디장이 자신의 스터디 페이지에서 신청 테이블을 읽어옴
@@ -131,7 +55,7 @@ public class StudyManageController {
      * memberId, studyId는 applyForStudyId에 해당하는 신청 테이블의 컬럼 값임(스터디장의 memberId값이 아님)
      * 우선 ScriptUtils를 이용하여 간단하게 로직 작성함(map에 오류를 넣어서 페이지로 반환시키든 리펙토링 필요)
      */
-    @PatchMapping("/studyManage/manager/approval/{studyId}/{memberId}/{applyForStudyId}")
+    @PatchMapping("/studyManage/manager/{studyId}/approval/{memberId}/{applyForStudyId}")
     public void approval(@PathVariable("studyId") Long studyId,
                            @PathVariable("memberId") Long memberId,
                            @PathVariable("applyForStudyId") Long applyForStudyId,
@@ -154,7 +78,7 @@ public class StudyManageController {
         }
     }
 
-    @GetMapping("/studyManage/manager/reject/{studyId}/{applyForStudyId}")
+    @GetMapping("/studyManage/manager/{studyId}/reject/{applyForStudyId}")
     @ResponseBody
     public ResponseEntity<?> isRejectable(@PathVariable("applyForStudyId") Long applyForStudyId) {
 
@@ -170,7 +94,7 @@ public class StudyManageController {
      * 스터디장이 신청 테이블의 목록에서 거절하는 로직
      * 스터디장이 거절했을 시 사유 테이블에 거절 메시지 작성
      */
-    @PostMapping("/studyManage/manager/reject/{studyId}/{applyForStudyId}")
+    @PostMapping("/studyManage/manager/{studyId}/reject/{applyForStudyId}")
     public void reject(@PathVariable("applyForStudyId") Long applyForStudyId,
                        @NotBlank @Range(min = 10, max = 500) @ModelAttribute("rejectReason")
                        String rejectReason,
@@ -198,7 +122,7 @@ public class StudyManageController {
     /**
      * 스터디장의 스터디 현황 페이지에서 현재 속해있는 스터디의 구성원을 강퇴시키는 기능
      */
-    @PatchMapping("/studyManage/manager/kickout/{studyId}/{studyMemberId}")
+    @PatchMapping("/studyManage/manager/{studyId}/kickout/{studyMemberId}")
     public void kickOut(@PathVariable("studyId") Long studyId,
                           @PathVariable("studyMemberId") Long studyMemberId,
                           HttpServletResponse response,
@@ -234,58 +158,6 @@ public class StudyManageController {
         // 해당 studyMember의 AttendState를 '강퇴'로 바꾸는 로직
         if (canKickOut)
             studyMemberService.changeAttendState(studyMemberId, AttendState.강퇴);
-
     }
 
-    private boolean validMessage(String appealMessage,Model model) {
-        if (appealMessage.length()< MIN_APPEALMESSAGE_LENGTH) {
-            model.addAttribute("invalidMessageLength", true);
-            return false;
-        }
-
-        if (appealMessage.length()> MAX_APPEALMESSAGE_LENGTH) {
-            model.addAttribute("invalidMessageLength", true);
-            model.addAttribute("rejectedMessage", appealMessage.substring(0,MAX_APPEALMESSAGE_LENGTH-1));
-            return false;
-        }
-        return true;
-    }
-
-    private boolean validateMemberAttendState(Long studyId, Long loginMemberId, HttpServletResponse response) throws IOException {
-
-        if (!studyMemberService.doesMemberBelongToStudy(studyId, loginMemberId)) {
-
-          return true;
-
-        } else {
-            AttendState attendState = studyMemberService.getAttendState(studyId, loginMemberId);
-
-            switch (attendState) {
-                case 참여 : {ScriptUtils.alertAndBackPage(response,"이미 가입하신 스터디입니다."); break; }
-                case 탈퇴 : {ScriptUtils.alertAndBackPage(response,"탈퇴한 스터디는 가입이 불가능합니다."); break; }
-                case 강퇴 : {ScriptUtils.alertAndBackPage(response,"강퇴당한 스터디는 가입이 불가능합니다."); break; }
-            }
-
-            return false;
-        }
-    }
-
-    private boolean validateBelongedStudyCount(Long loginMemberId) throws IOException {
-        int studyCount = studyMemberService.getBelongedStudyCount(loginMemberId);
-
-        if(studyCount>=STUDY_COUNT_MAX) {
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * 이미 신청한 상태인지 확인하는 로직
-     * @param studyId
-     * @param loginMemberId
-     */
-    private boolean validateNotEnrolled(Long studyId, Long loginMemberId) {
-        return applyForStudyService.findApplyForStudyByMemberAndStudy(studyId, loginMemberId).isEmpty();
-    }
 }
