+++ src/main/resources/templates/boards/board.html
@@ -664,7 +664,7 @@
     applyForStudy = function() {
         var r = confirm("해당 스터디에 가입하시겠습니까?");
         if (r == true) {
-            location.href = "/studyManage/enroll/" + [[${studyId}]];
+            location.href += "/enroll";
         }
     }
